Move 3rd party libs to static library.
