+++ lib/include/khrplatform.h
