+++ lib/src/glad.cpp
@@ -1,3 +1,5 @@
+extern "C" {
+
 /*
 
     OpenGL loader generated by glad 0.1.36 on Fri Feb 14 15:17:34 2025.
@@ -640,7 +642,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <lib/glad/glad.h>
+#include <glad.h>
 
 static void* get_proc(const char *namez);
 
@@ -9431,3 +9433,4 @@ int gladLoadGLLoader(GLADloadproc load) {
 	return GLVersion.major != 0 || GLVersion.minor != 0;
 }
 
+}
